+++ src/Components/Dialog/DialogReference.DialogWithResult.cs
@@ -1,34 +1,39 @@
-﻿using System.Diagnostics.CodeAnalysis;
-using BlazorUiKit.Abstractions.Dialog;
-
-namespace BlazorUiKit.Components;
+﻿namespace BlazorUiKit.Components;
 
 internal sealed class DialogReference<[DynamicallyAccessedMembers(DynamicallyAccessedMemberTypes.All)] TDialog, TResult> : DialogReferenceBase, IDialogReference<TDialog, TResult>
-	where TDialog : DialogBase<TDialog, TResult>
+    where TDialog : DialogBase<TDialog, TResult>
 {
-	public DialogReference()
-	{
-		InitializeContentRenderFragment<TDialog>();
-	}
+    public DialogReference()
+    {
+        InitializeContentRenderFragment<TDialog>();
+    }
 
-	public Task<TResult> CompletionTask => _tcs.Task;
+    public Task<TResult> CompletionTask => _tcs.Task;
 
-	private readonly TaskCompletionSource<TResult> _tcs = new();
+    private readonly TaskCompletionSource<TResult> _tcs = new();
 
-	public override void Cancel()
-	{
-		_tcs.TrySetCanceled();
-		DialogService.RemoveDialog(Id);
-	}
+    public override void Cancel()
+    {
+        if (IsCanceledOrClosed)
+            return;
 
-	public void Close(TResult result)
-	{
-		if (ActualDialog?.OnClosing() == false)
-		{
-			return;
-		}
+        IsCanceledOrClosed = true;
+        _tcs.SetCanceled();
+        DialogService.RemoveDialog(Id);
+    }
 
-		DialogService.RemoveDialog(Id);
-		_tcs.TrySetResult(result);
-	}
+    public void Close(TResult result)
+    {
+        if (IsCanceledOrClosed)
+            return;
+
+        ActualDialog?.OnClosing();
+
+        if (ActualDialog?.CanClose != true)
+            return;
+
+        _tcs.SetResult(result);
+        DialogService.RemoveDialog(Id);
+        IsCanceledOrClosed = true;
+    }
 }
